changed knockback resistance to a float from a int

+++ src/main/java/com/moltenwolfcub/playtime/armor/ModArmorMaterial.java
@@ -16,7 +16,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 public enum ModArmorMaterial implements ArmorMaterial{
 
     FOAM(Playtime.MOD_ID + ":foam", 8, new int[] { 1, 2, 3, 1 }, 5,
-        SoundEvents.ARMOR_EQUIP_GENERIC, 0.0F, () -> { return Ingredient.of(RegistryHandler.FOAM.get()); }, 0);
+        SoundEvents.ARMOR_EQUIP_GENERIC, 0.0F, () -> { return Ingredient.of(RegistryHandler.FOAM.get()); }, 0.0f);
 
     private static final int[] MAX_DAMAGE_ARRAY = new int[] { 11, 16, 15, 13 };
     private final String name;
@@ -26,10 +26,10 @@ public enum ModArmorMaterial implements ArmorMaterial{
     private final SoundEvent soundEvent;
     private final float toughness;
     private final Supplier<Ingredient> repairMaterial;
-    private final int knockbackResistance;
+    private final float knockbackResistance;
 
     ModArmorMaterial(String name, int maxDamageFactor, int[] damageReductionAmountArray, int enchantability,
-     SoundEvent soundEvent, float toughness, Supplier<Ingredient> repairMaterial, int knockbackResistance) {
+     SoundEvent soundEvent, float toughness, Supplier<Ingredient> repairMaterial, float knockbackResistance) {
          this.name = name;
          this.maxDamageFactor = maxDamageFactor;
          this.damageReductionAmountArray = damageReductionAmountArray;
